Updated SourceInputProvider and tests
It would be nice if the codebase was a tad more "discoverable" - it should be simple to get a list of capabilities. In that spirit, this is an effort with SourceInputProvider to generate all currently "baked-in" source inputs (previously only returned three types). Was working on a more generic approach (see commented Activator.CreateInstance and Type.GetConstructor calls), but was getting runtime exceptions for both...not sure what's up there. The ultimate goal would be to report all supported inputs, even those made available by the consumer.